requirements.txt
mysql-connector
flask
mysqlclient
requests
slackclient
mysql-connector-python
datetime
gunicorn
